Refactored components/PlayFriendMode/Buttons into components/PlayFriendMode/GameButtons
